+++ Makefile.am
@@ -52,7 +52,12 @@ memcached_SOURCES += sasl_defs.c
 endif
 
 if ENABLE_PROXY
-memcached_SOURCES += proto_proxy.c proto_proxy.h vendor/mcmc/mcmc.h
+memcached_SOURCES += proto_proxy.c proto_proxy.h vendor/mcmc/mcmc.h \
+					 proxy_xxhash.c proxy.h \
+					 proxy_await.c proxy_ustats.c \
+					 proxy_jump_hash.c proxy_request.c \
+					 proxy_network.c proxy_lua.c \
+					 proxy_config.c
 endif
 
 if ENABLE_EXTSTORE
